+++ client/src/Pages/registration/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
-import LogIn from '../../Components/Login'
-import SignUp from '../../Components/SignUp'
+import LogIn from '../../Components/login'
+import SignUp from '../../Components/signUp'
 import { ReactComponent as Logo } from '../../imgs/shellhacks.svg'
 import './styles.css'
